Unidiff now retains more infromation found in diffs.
Lines containing '@@ -1,2 3,4 @@ This line of text' are now handled correctly.

The original sources were at version 1.1 so update the VERSION accordingly.

Clarification form the author of unidff as to it's license.

Sources are now hosted at https://github.com/AceHusky12


Ok Wayne Davison, the inventor of the unified diff and the author of unidiff.