x11/wxGTK32: import wxGTK32-3.2.0
Packaged with help from dbj@

wxWidgets is a set of libraries that allows C++ applications to compile and
run on several different types of computer, with minimal source code changes.
There is one library per supported GUI (such as Windows, GTK+, Motif, and Mac).
As well as providing a common API (Application Programming Interface) for GUI
functionality, it provides functionality for accessing some commonly-used
operating system facilities, from copying and deleting files to socket and
thread support. wxWidgets is a 'framework' in the sense that it provides a lot
of built-in functionality, which the application can use or replace as required,
thus saving a great deal of coding effort. Basic data structures such as
strings, arrays, linked lists and hash tables are also supported.

Such a toolkit is not unique - there are others to choose from - but wxWidgets
is free, well-established, well-documented, and very broad in its coverage of
GUI functionality. It has some extras that make it stand out from the crowd,
such as the many convenience dialogs, built-in HTML display and printing,
virtual filesystems, easy-to-use OLE automation controller class, Open GL
support, and many other features that make it easier to write modern and
user-friendly applications.

This package contains the 3.2 stable branch of the library.