lintpkgsrc: cleanup: merge 'store' subroutines
With small changes, the subroutine for storing the cache can be
restricted to the public API of the types PkgDb, Pkgs and PkgVer. This
way, the 'store' subroutines no longer need to be part of the classes.
Implementing the code in a single external subroutine saves a few lines
of code.

The order of the PkgVer entries in the cache file changes, everything
else is as before.