fonts/tex-amiri: import tex-amiri-0.113
Amiri is a classical Arabic typeface in Naskh style for
typesetting books and other running text. It is a revival of
the beautiful typeface pioneered in the early 20th century by
Bulaq Press in Cairo, also known as Amiria Press, after which
the font is named. The project aims at the revival of the
aesthetics and traditions of Arabic typesetting, and adapting
it to the era of digital typesetting, in a publicly available
form.